+++ front/src/Socket.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import toast from 'react-hot-toast';
 import { SetterOrUpdater, useSetRecoilState } from 'recoil';
-import { IChartItem } from '@src/types';
+import { IChartItem, IDailyLog } from '@src/types';
 import chartAtom from '@recoil/chart';
 import HoldStockListAtom from '@recoil/holdStockList';
+import dailyLogAtom from '@recoil/stockDailyLog';
 import fetchHoldStocks from '@common/utils/fetchHoldStocks';
 import webSocketAtom from '@recoil/websocket/atom';
 import stockListAtom, { IStockListItem, IStockChartItem } from '@recoil/stockList/atom';
@@ -11,7 +12,6 @@ import { IAskOrderItem, IBidOrderItem, askOrdersAtom, bidOrdersAtom } from '@rec
 import stockExecutionAtom, { IStockExecutionInfo, IStockExecutionItem } from './recoil/stockExecution/atom';
 import { translateRequestData, translateResponseData } from './common/utils/socketUtils';
 import Emitter from './common/utils/eventEmitter';
-import dailyLogAtom, { IDailyLog } from './recoil/stockDailyLog/atom';
 
 interface IProps {
 	children: React.ReactNode;
