+++ front/src/recoil/stockDailyLog/index.ts
@@ -1,5 +1,1 @@
-import atom from './atom';
-import type { IDailyLog } from './atom';
-
-export { IDailyLog };
-export default atom;
+export { default } from './atom';
